Separate asset config in moonbeam runtime
